requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.15,<3.16
django-cors-headers>=4.4,<4.5
whitenoise>=6.7,<6.8
gunicorn>=22.0,<23.0
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.2,<2.3
python-dotenv>=1.0,<1.1
